Worked through several issues with GitHub API
I got authenticated access working using
a OAuth token stored in the environment
variable `GITHUB_TOKEN`.

But I ran into issues with access the
`xc_results.json` file because it is so
large and the GitHub API doesn't like you
to download large files so they redirect
you to the Data API.  So now all my
read access is via the Data API.

However, then I ran into request throttling
where GitHub only lets you perform a
small number of **unauthenticated**
requests over time.  So I modified the
reading section to leverage API keys
if they are available (side stepping
some of the throttling).

If the `GITHUB_TOKEN` is not provided
the script will warn the user that the
data has been processed, but not
uploaded.