Using vendorId
This uses the newer version of `fmi-data`
(version 0.8.0+) that defines a `vendorId`
field in both the tools and vendor data.
This replaces the `repouri` as a way to
"join" these two data sources.

We also now parse the `vendor.ini` file
to extract the `vendorId` so it doesn't
need to be specified on the command
line.  The directory is also now `.` by
default.  So normally, no command line
arguments are required at this point
to run `process_repo`.

This fixes #3.