Major overhaul, code compiles again
I've made some huge changes to the database API (with more to come possibly).  I got
rid of existing Mongo and Github database implementations because they will have
to be rewritten from scratch (if at all).  This leverages fmi-data v0.9.0 as well
which adds a vendorId flag to all entities in the database.