Writing to test repo
This seems to skip commits if no
changes are made.  So, this fixes #5.

Also, this implements a new Github
backend that doesn't use the Github
API at all (thus avoiding all that large
file stuff).  So, this fixes #10.

Finally, I checked and it seems as though
this doesn't generate canonical output
as dictate by #8.

Now I'm going to run some tests with
CI before changing the defaults to point
to the real fmi-standard.org repo.