Compiling and running again
Lots of reorganization fo code.  Trying to emphasizing readability so I can read this
after I get back from the holidays and so others can understand and modify it in
the future.

I recognize that for initializing the repos, I should do a bunch of up front work
to cache all import information since I'm rereading it with every tool.

Copying files is currently commented out, I need to put that back.  But it is
turned off for now to speed up debugging.